requirements.txt
pysstv
Pillow
pydub
